Sensor API.
Implemented a simple category sensor that selects random fields for each
category.

The API involves a RawInput channel, where each input is named and has a
value (currently only int supported); The RawInput's value is passed to
a decoder function that creates the bitset, which is in turn merged back
into an Input struct with the same label as the raw input.