Initial learning implementation.
This is still very crude. I believe a logistic function to update the
permanence would make more sense here than using a fixed increment.

Numenta swarms over (i.e. trains a model for) the increment values, but
I think the fixed increments brings many problems. In particular,
numbing a column in a way that causes it to never activate again is bad.
They induce a reset for those cases, which forgets recent history.