A few performance optimization attempts:
1. Improved AdaptSegments().

We only need to look at the predicted cell from the previous step.

Before:
BenchmarkTp_AAB	   10000	    267453 ns/op

After:
BenchmarkTp_AAB	   10000	    238486 ns/op

2. Improved FindBestSegment().

Early return once we find a segment with the maximum possible overlap.

3. Benchmarks for NumSetBits.

I was trying to optimize it to a shifting magic function, but it turns out that
this version is better for few set bits, which is the common case for HTM.

So we will not change it even though it is worse for denser bitsets.